End of Week Progress
To Do List Progress.

Struggled to work in styles on events. Went back to codeacademy to work on basics and got up to working with more complex strings and inserting if/then statements in the proper syntax. Struggled to get styled components running, and the external stylesheet made it difficult to manage events. React was difficult to grasp for me personally.